MunuListServlet now gets the data from datastore, then send to front end
