--- src/main/java/com/google/sps/servlets/ListTasksServlet.java
@@ -1,63 +1,0 @@
-// Copyright 2019 Google LLC
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     https://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package com.google.sps.servlets;
-
-import com.google.cloud.datastore.Datastore;
-import com.google.cloud.datastore.DatastoreOptions;
-import com.google.cloud.datastore.Entity;
-import com.google.cloud.datastore.Query;
-import com.google.cloud.datastore.QueryResults;
-import com.google.cloud.datastore.StructuredQuery.OrderBy;
-import com.google.gson.Gson;
-// import com.google.sps.data.MenuItem;
-import java.io.IOException;
-import java.util.ArrayList;
-import java.util.List;
-import javax.servlet.annotation.WebServlet;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-/** Servlet responsible for listing tasks. */
-@WebServlet("/list-tasks")
-public class ListTasksServlet extends HttpServlet {
-
-  @Override
-  public void doGet(HttpServletRequest request, HttpServletResponse response) throws IOException {
-    Datastore datastore = DatastoreOptions.getDefaultInstance().getService();
-    Query<Entity> query =
-        Query.newEntityQueryBuilder().setKind("MenuItem").setOrderBy(OrderBy.desc("timestamp")).build();
-    QueryResults<Entity> results = datastore.run(query);
-    ArrayList<String> menulist = new ArrayList<String> ();
-    
-    while (results.hasNext()) {
-      Entity entity = results.next();
-
-      String id = String.valueOf(entity.getKey().getId());
-      String itemName = entity.getString("itemName");
-      String itemPrice = entity.getString("itemPrice");
-
-      menulist.add(id);
-      menulist.add(itemName);
-      menulist.add(itemPrice);
-   
-    }
-
-    Gson gson = new Gson();
-
-    response.setContentType("application/json;");
-    response.getWriter().println(gson.toJson(menulist));
-  }
-}
